drf-auth/requirements.txt
asgiref==3.4.1; python_version >= "3.6"
django==3.2.10; python_version >= "3.6"
djangorestframework-simplejwt==5.0.0; python_version >= "3.7"
djangorestframework==3.13.0; python_version >= "3.6"
gunicorn==20.1.0; python_version >= "3.5"
psycopg2-binary==2.9.2; python_version >= "3.6"
pyjwt==2.3.0; python_version >= "3.7"
pytz==2021.3; python_version >= "3.7"
sqlparse==0.4.2; python_version >= "3.6"
whitenoise==5.3.0; python_version >= "3.5" and python_version < "4"
